Direction of photon counting changed between frames
***this is a major change***
A fourth counter is implemented in kinetic series to switch the
direction of photon counting between frames.

It turns out that the edge counter of the NI X-Series boards have a
readout register of 24 bits.  This produces the unfortunate artifact of
rounding photon counts to the nearest bit once the total number of
counts passes 16 million.  To avoid hitting 16 million I start the
counting at 8 million and switch the direction of counting between
frames. (i.e. Frame 0 counts away from 0 and frame 1 counts towards 0).